Allow checking for constant values in memory
We forbid insertions of constant values into the memory, as
then get() will return constant values. If these values are
written to the state we'll get an error for writing a constant
value to the trace, which is a false positive.

However, when a (non-constant) value already exists at a
given address, it is OK for set() to compare it to a constant
value. This is useful in run mode when an AirFn does a set()
with a constant value.