Change opcode tests to use literal state & air_body values
Prior to this commit the tests contained logic to compute the
expected state and air_body for every possible instance_def.
It will be hard to maintain this logic as the air_body for this
opcode becomes more complex (e.g., due to the introduction of
Felt252IdMemory).

The current design makes it easier to adapt the test when the
implementation of the opcode (or one of its dependencies) changes.