Bugfix: CallOpcode: Verify low 27-bits of saved fp and pc, not just 9
CallOpcode verifies that [ap] = fp and [ap+1] = return_pc. These
checks were done using a MemVerify call that assumed that fp and
return_pc fit in a single limb (9 bits).

Replace these with memory.read_address that allows fp and
return_pc to be up to 27 bits wide.

In addition, have a better description for the state cells that
contain the distance to the next pc in the relative-call case.