Remove incorrect assertion in AirFnRegistry::build_air
LookupData (today named LookupTerm) contains the individual felts
of the value to be yielded / used. That value is guaranteed to be
in-state by the assert below, so the felts will be state cell
references and the value itself won't be accessed by the constraint
evaluation code.

Also, add some comments.