Felts that are directly in state don't need parents
For felts that are directly_in_state, the parent value is not used when compiling (the felt is accessed as state[something], not parent.something). Previously we were maintaining their parent value to have more readable JSONs for the visualizer.
Now it was decided that the visualizer will work with the compiled JSONs (which don't contain parent information), and therefore maintaining the parent of directly_in_state felts is no longer required.