Fix variable names in constraints and add input/output descriptions
Prior to this commit, for single-felt intermediates, the same name was
used in both constraints and deductions. This name was independent of
the input_expr_descriptions.

This commit separates the deduction name and constraints name in the
output of rec_let (two different AirBodyComponent::Intermediate are
generated). This way the input_expr_descriptions influence the names
of single-felt variables too (in constraints) which allows better
visualizer output.