Use a list for the names of intermediates created by a StaticCall
Until now these were stored as a string in the `name` field
(e.g., "[felt_1, felt_2]") of the `Intermediate` constraint step,
which made it more complicated to parse by automated tools.

The change is only in the "constraints" part of the JSON. The
"deductions" field is left as-is as StaticCalls there always create
a single intermediate (which might be of some complex type).