Don't compile code_gen_regression
code_gen_regression contains the generated code so that the effect
of changes to the generator is visible in `git diff`. We used to
also compile the generated code (using mocks for various parts of
stwo-cairo), to test that the generator outputs valid code.

Now that we have a test that compiles the generated code inside
stwo-cairo (test_generated_code.py) we don't need to compile
code_gen_regression anymore and don't need the mocks either.

In addition, because the code_gen_regression directory is now only
for human reviewers and not for automated tests, we removed
files that are nearly identical and files for components we don't
plan to auto-generate in the near future (memory).